Refactor update settings dialog and GitHub token management. Moved the rendering of the update settings dialog to the updater class for better organization. Enhanced the UI for version selection and GitHub token input, including improved token validation and error handling. Removed redundant methods from the GUI class to streamline the codebase.
